Stop using C-extension functions for parsing datetime.
Accept only ISO8601 format for the `datetime` and `local_datetime`
casts from str (or json) into the respective datetime types. The
check and cast are now done in pure SQL.

Add a hint to the cast format error that explains that ISO8601 is
expected and that converter functions can be used for additional
formatting options.

Issue #740